I changed all the names of the txtBoxes and worked on the design of the class diagram. Along with making things cleaner and clearing up some code.
