+++ KeithsUI_Feb14/BunchOfMoney.cs
@@ -3,7 +3,7 @@ using System.Collections.Generic;
 using System.Linq;
 using System.Text;
 
-namespace Vending_Machince
+namespace VendingMachine
 {
     public class BunchOfMoney
     {
