+++ KeithsUI_Feb14/ClassDiagram1.cd
@@ -1,113 +1,120 @@
 ﻿<?xml version="1.0" encoding="utf-8"?>
 <ClassDiagram MajorVersion="1" MinorVersion="1">
-  <Class Name="Vending_Machince.Class1" Collapsed="true">
-    <Position X="5.25" Y="6.75" Width="1.5" />
+  <Class Name="VendingMachine.Class1">
+    <Position X="1.75" Y="5.25" Width="1.5" />
     <TypeIdentifier>
       <HashCode>AAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAA=</HashCode>
       <FileName>Class1.cs</FileName>
     </TypeIdentifier>
   </Class>
-  <Class Name="Vending_Machince.Form1" Collapsed="true">
-    <Position X="7" Y="6.25" Width="1.5" />
+  <Class Name="VendingMachine.VendingMachinefrm" Collapsed="true">
+    <Position X="7.75" Y="4.5" Width="1.5" />
     <TypeIdentifier>
-      <HashCode>ACAAAAAAACAAABABAACAAAgCAMwrAAAEEHIMCAAAAAA=</HashCode>
-      <FileName>Form1.cs</FileName>
+      <HashCode>AABAQAAACCEAABAAQAjAAAAGACAAAAAAEAAAAABAAgA=</HashCode>
+      <FileName>VendingMachinefrm.cs</FileName>
     </TypeIdentifier>
   </Class>
-  <Class Name="Vending_Machince.Program" Collapsed="true">
-    <Position X="7" Y="5.25" Width="1.5" />
+  <Class Name="VendingMachine.Program" Collapsed="true">
+    <Position X="1.75" Y="4.5" Width="1.5" />
     <TypeIdentifier>
       <HashCode>AAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAABAAAAAA=</HashCode>
       <FileName>Program.cs</FileName>
     </TypeIdentifier>
   </Class>
-  <Class Name="Vending_Machince.Properties.Resources" Collapsed="true">
-    <Position X="5.25" Y="7.5" Width="1.5" />
+  <Class Name="VendingMachine.Properties.Resources" Collapsed="true">
+    <Position X="3.5" Y="5.25" Width="1.5" />
     <TypeIdentifier>
       <HashCode>AAAAAAAAAAAAAAAAAAABFAAAAQAAAAAAAAAAAAAAAIA=</HashCode>
     </TypeIdentifier>
   </Class>
-  <Class Name="Vending_Machince.Properties.Settings" Collapsed="true">
-    <Position X="7" Y="7.25" Width="1.5" />
+  <Class Name="VendingMachine.Properties.Settings" Collapsed="true">
+    <Position X="3.5" Y="4.5" Width="1.5" />
     <TypeIdentifier>
       <HashCode>AAAAAAAAAAAAAAAAAAAAIAAAAAABAAAAAAAAAAAAAAA=</HashCode>
     </TypeIdentifier>
   </Class>
-  <Class Name="Vending_Machince.vending_machine">
-    <Position X="8.75" Y="5" Width="1.5" />
+  <Class Name="VendingMachine.VendingMachine">
+    <Position X="7.75" Y="7.5" Width="1.5" />
     <TypeIdentifier>
-      <HashCode>AAAAAAAAAAAAAAAAEAQAAAAAAAAAAAAAAAAAAAAAAAg=</HashCode>
-      <FileName>vending_machine.cs</FileName>
+      <HashCode>AAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAg=</HashCode>
+      <FileName>VendingMachine.cs</FileName>
     </TypeIdentifier>
   </Class>
-  <Class Name="Vending_Machince.Product_Storage">
-    <Position X="10.75" Y="6.5" Width="1.5" />
+  <Class Name="VendingMachine.ProductStorage">
+    <Position X="11.25" Y="5.5" Width="1.5" />
     <TypeIdentifier>
       <HashCode>AAAAAAQAAAAAAAAAAAAIAAAAAAAAAAAAAAAAAAAAAAA=</HashCode>
       <FileName>Product_Storage.cs</FileName>
     </TypeIdentifier>
   </Class>
-  <Class Name="Vending_Machince.Products">
-    <Position X="8.75" Y="7.25" Width="1.5" />
+  <Class Name="VendingMachine.Products">
+    <Position X="9.5" Y="5.5" Width="1.5" />
     <TypeIdentifier>
-      <HashCode>AAABAAAAAAAAAAAAAAAAAAAAAAAAAAACAAAAAAAAAAA=</HashCode>
+      <HashCode>AAAAAAAAAQAAAAAAAAAAAAAAAAAAAAAAAAAAAgAAAAA=</HashCode>
       <FileName>Products.cs</FileName>
     </TypeIdentifier>
   </Class>
-  <Class Name="Vending_Machince.Displays">
-    <Position X="12.25" Y="5.25" Width="1.5" />
+  <Class Name="VendingMachine.Display">
+    <Position X="11.25" Y="7" Width="1.5" />
     <TypeIdentifier>
       <HashCode>AAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAA=</HashCode>
       <FileName>Displays.cs</FileName>
     </TypeIdentifier>
   </Class>
-  <Class Name="Vending_Machince.CashStorage">
-    <Position X="12.25" Y="9.25" Width="2.25" />
+  <Class Name="VendingMachine.CashStorage">
+    <Position X="9.5" Y="9.25" Width="2.25" />
     <TypeIdentifier>
       <HashCode>AAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAA=</HashCode>
       <FileName>CashStorage.cs</FileName>
     </TypeIdentifier>
   </Class>
-  <Class Name="Vending_Machince.Bank">
-    <Position X="5.25" Y="5.25" Width="1.5" />
+  <Class Name="VendingMachine.Bank">
+    <Position X="3.5" Y="9.25" Width="1.5" />
     <TypeIdentifier>
       <HashCode>AAAAAAAAAAAAAAAAAAAAAAAAAAIAAAAAAAAAAAAQAAA=</HashCode>
       <FileName>Bank.cs</FileName>
     </TypeIdentifier>
   </Class>
-  <Class Name="Vending_Machince.Customer">
-    <Position X="5.25" Y="8.25" Width="1.5" />
-    <TypeIdentifier>
-      <HashCode>AABQEAAAAAAgggAAQAAAAAAAAgAAAAAAAAgAAAACAAA=</HashCode>
-      <FileName>Customer.cs</FileName>
-    </TypeIdentifier>
-  </Class>
-  <Class Name="Vending_Machince.WorldOfMoney">
-    <Position X="7" Y="8.25" Width="1.5" />
+  <Class Name="VendingMachine.WorldOfMoney">
+    <Position X="3.5" Y="6" Width="1.5" />
     <TypeIdentifier>
       <HashCode>AAAAAAAAAAAAAAAAAAABAAAAAAYACAAAAAAAAAAAAAA=</HashCode>
       <FileName>WorldOfMoney.cs</FileName>
     </TypeIdentifier>
   </Class>
-  <Class Name="Vending_Machince.BunchOfMoney">
+  <Class Name="VendingMachine.BunchOfMoney">
     <Position X="3.5" Y="8.25" Width="1.5" />
     <TypeIdentifier>
       <HashCode>AAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAA=</HashCode>
       <FileName>BunchOfMoney.cs</FileName>
     </TypeIdentifier>
   </Class>
-  <Class Name="Vending_Machince.Storage">
-    <Position X="13" Y="6.5" Width="1.5" />
+  <Class Name="VendingMachine.Storage">
+    <Position X="9.5" Y="7" Width="1.5" />
     <TypeIdentifier>
-      <HashCode>AAAAAAIAAAABAAAAAAAAAAAAAAAAAAAAAAABAAAAAAA=</HashCode>
+      <HashCode>AAAAAAIAAAAEAAAAAAAAAAAAAAAAAAAAAAABAAAAAAA=</HashCode>
       <FileName>Storage.cs</FileName>
     </TypeIdentifier>
   </Class>
-  <Class Name="Vending_Machince.InputCash">
-    <Position X="9.5" Y="9.25" Width="1.5" />
+  <Class Name="VendingMachine.InputCash">
+    <Position X="7.75" Y="9.25" Width="1.5" />
     <TypeIdentifier>
       <HashCode>AAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAA=</HashCode>
       <FileName>InputCash.cs</FileName>
+    </TypeIdentifier>
+  </Class>
+  <Class Name="VendingMachine.Customer">
+    <Position X="0.5" Y="6.25" Width="2.75" />
+    <TypeIdentifier>
+      <HashCode>CAAQEAAAAAAAAgAAQAAAAAAAAgAAAAAAAAAAAAIAAAA=</HashCode>
+      <FileName>Customer.cs</FileName>
+    </TypeIdentifier>
+  </Class>
+  <Class Name="VendingMachine.Vendor">
+    <Position X="7.75" Y="5.5" Width="1.5" />
+    <TypeIdentifier>
+      <HashCode>AAAAAAAAAAAEAAAAEAQAAAAAAAAAAAAAAAAAAAAAAAA=</HashCode>
+      <FileName>Vendor.cs</FileName>
     </TypeIdentifier>
   </Class>
   <Font Name="Segoe UI" Size="9" />
