+++ KeithsUI_Feb14/Product_Storage.cs
@@ -3,10 +3,10 @@ using System.Collections.Generic;
 using System.Linq;
 using System.Text;
 
-namespace Vending_Machince
+namespace VendingMachine
 {
 
-    public class Product_Storage
+    public class ProductStorage
     {
 
         public void RemoveProducts()
