+++ KeithsUI_Feb14/Vending Machine.csproj
@@ -55,11 +55,11 @@
     <Compile Include="Class1.cs" />
     <Compile Include="Customer.cs" />
     <Compile Include="Displays.cs" />
-    <Compile Include="Form1.cs">
+    <Compile Include="VendingMachinefrm.cs">
       <SubType>Form</SubType>
     </Compile>
-    <Compile Include="Form1.Designer.cs">
-      <DependentUpon>Form1.cs</DependentUpon>
+    <Compile Include="VendingMachinefrm.Designer.cs">
+      <DependentUpon>VendingMachinefrm.cs</DependentUpon>
       <CopyToOutputDirectory>PreserveNewest</CopyToOutputDirectory>
     </Compile>
     <Compile Include="InputCash.cs" />
@@ -68,10 +68,11 @@
     <Compile Include="Program.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
     <Compile Include="Storage.cs" />
-    <Compile Include="vending_machine.cs" />
+    <Compile Include="VendingMachine.cs" />
+    <Compile Include="Vendor.cs" />
     <Compile Include="WorldOfMoney.cs" />
-    <EmbeddedResource Include="Form1.resx">
-      <DependentUpon>Form1.cs</DependentUpon>
+    <EmbeddedResource Include="VendingMachinefrm.resx">
+      <DependentUpon>VendingMachinefrm.cs</DependentUpon>
     </EmbeddedResource>
     <EmbeddedResource Include="Properties\Resources.resx">
       <Generator>ResXFileCodeGenerator</Generator>
