+++ KeithsUI_Feb14/VendingMachine.cs
@@ -1,0 +1,15 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+
+namespace VendingMachine
+{
+    public class VendingMachine : Vendor
+    {
+        public void calculateChange()
+        {
+            throw new System.NotImplementedException();
+        }
+    }
+}
