+++ KeithsUI_Feb14/VendingMachinefrm.cs
@@ -1,0 +1,17 @@
+﻿using System;
+using System.Collections.Generic;
+using System.ComponentModel;
+using System.Data;
+using System.Drawing;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+using System.Windows.Forms;
+using MoneyLib;
+namespace VendingMachine
+{
+    public partial class VendingMachinefrm : Form
+    {
+       
+    }
+}
