+++ KeithsUI_Feb14/Vendor.cs
@@ -1,0 +1,22 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+
+namespace VendingMachine
+{
+    public class Vendor
+    {
+        private int listOfMoney;
+
+        public void InputMoney()
+        {
+            throw new System.NotImplementedException();
+        }
+
+        public void TakeMoney()
+        {
+            throw new System.NotImplementedException();
+        }
+    }
+}
