--- KeithsUI_Feb14/vending_machine.cs
@@ -1,25 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-
-namespace Vending_Machince
-{
-    public class vending_machine
-    {
-        public void calculateChange()
-        {
-            throw new System.NotImplementedException();
-        }
-
-        public void TakeMoney()
-        {
-            throw new System.NotImplementedException();
-        }
-
-        public void InputMoney()
-        {
-            throw new System.NotImplementedException();
-        }
-    }
-}
